refactor: torna WithSorting reutilizável
